# More optimized remainder calculation