Instead of division then remainder operation, used Math.DivRem which handles it in a more optimized way. It uses only a single division, multiplication, and subtraction instead of two divisions. Moreover, in the future, it can be even more optimized using intrinsics.

See:
https://github.com/dotnet/runtime/issues/5213#issuecomment-260448104