@paulomorgado 2023-02-13
It would be nice if this PR was acompanied by benchmarks.