@lechu445 2023-02-13
before:
```
|    Method |     Mean |   Error |   StdDev |
|---------- |---------:|--------:|---------:|
| WriteByte | 184.0 ms | 3.64 ms |  5.34 ms |
| WriteSpan | 366.3 ms | 7.07 ms | 12.00 ms |
```

after:
```
|    Method |     Mean |   Error |   StdDev |
|---------- |---------:|--------:|---------:|
| WriteByte | 115.5 ms | 2.30 ms |  3.23 ms |
| WriteSpan | 319.7 ms | 6.28 ms | 11.49 ms | 
```

code:
```c#
public class Test
{
    private RecyclableMemoryStream subject;
    private readonly byte[] bytes = new byte[] { 1, 2, 3, 4, 5, 6 };

    [IterationSetup]
    public void IterationSetup()
    {
        var manager = new RecyclableMemoryStreamManager();
        this.subject = new RecyclableMemoryStream(manager);
    }

    [Benchmark]
    public void WriteByte()
    {
        for (int i = 0; i < 10_000_000; i++)
        {
            this.subject.WriteByte(1);
        }
    }

    [Benchmark]
    public void WriteSpan()
    {
        for (int i = 0; i < 10_000_000; i++)
        {
            this.subject.Write(bytes.AsSpan());
        }
    }
}
```