@paulomorgado 2023-02-13
I was thinking more in the lines of:

```csharp
public class Benchmarks
{
	[Benchmark(Baseline = true)]
	public (long quocient, long remainder) DivAndMod() => (4000L / 1024L, 4000L % 1024L);

	[Benchmark]
	public (long quocient, long remainder) MathDivRem() => Math.DivRem(4000L, 1024L);
}
```

|     Method |      Mean |     Error |    StdDev | Ratio | RatioSD |
|----------- |----------:|----------:|----------:|------:|--------:|
|  DivAndMod | 0.2267 ns | 0.0092 ns | 0.0077 ns |  1.00 |    0.00 |
| MathDivRem | 0.0607 ns | 0.0221 ns | 0.0271 ns |  0.25 |    0.12 |